minor fixes

+++ src/Lab.sol
@@ -31,10 +31,7 @@ contract Lab {
 
     mapping(uint256 => uint256) internal warpDriveLevel;
 
-    function techUpgradeCost(
-        uint256 currentLevel,
-        S.ERC20s memory cost
-    ) internal pure returns (S.ERC20s memory) {
+    function techUpgradeCost(uint256 currentLevel, S.ERC20s memory cost) internal pure returns (S.ERC20s memory) {
         S.ERC20s memory _cost;
         _cost.steel = cost.steel * 2 ** currentLevel;
         _cost.quartz = cost.quartz * 2 ** currentLevel;
@@ -50,10 +47,7 @@ contract Lab {
         require(labLevel >= 1, "Level 1 Lab req");
     }
 
-    function beamTechnologyRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function beamTechnologyRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 1, "Level 1 Lab req");
         require(techs.energyInnovation >= 1, "Level 1 Energy Innovation req");
     }
@@ -62,19 +56,13 @@ contract Lab {
         require(labLevel >= 2, "Level 2 Lab req");
     }
 
-    function ionSystemsRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function ionSystemsRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 4, "Level 4 Lab req");
         require(techs.beamTechnology >= 5, "Level 5 Beam tech req");
         require(techs.energyInnovation >= 4, "Level 4 Energy Innovation  req");
     }
 
-    function plasmaEngineeringRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function plasmaEngineeringRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 4, "Level 4 Lab req");
         require(techs.beamTechnology >= 10, "Level 10 Beam Tech req");
         require(techs.energyInnovation >= 8, "Level 8 Energy Innovation  req");
@@ -85,49 +73,34 @@ contract Lab {
         require(labLevel >= 4, "Level 4 Lab req");
     }
 
-    function shieldTechRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function shieldTechRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 6, "Level 6 Lab req");
         require(techs.energyInnovation >= 6, "Level 6 Energy Innovation  req");
     }
 
-    function spacetimeWarpRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function spacetimeWarpRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 7, "Level 7 Lab req");
         require(techs.energyInnovation >= 5, "Level 5 Energy Innovation  req");
         require(techs.shieldTech >= 5, "Level 5 Shield Tech req");
     }
 
-    function combustiveDriveRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function combustiveDriveRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 1, "Level 1 Lab req");
         require(techs.energyInnovation >= 1, "Level 1 Energy Innovation  req");
     }
 
-    function thrustPropulsionRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function thrustPropulsionRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 2, "Level 2 Lab req");
         require(techs.energyInnovation >= 1, "Level 1 Energy Innovation  req");
     }
 
-    function warpDriveRequirements(
-        uint256 labLevel,
-        S.Techs memory techs
-    ) internal pure {
+    function warpDriveRequirements(uint256 labLevel, S.Techs memory techs) internal pure {
         require(labLevel >= 7, "Level 7 Lab req");
         require(techs.energyInnovation >= 5, "Level 5 Energy Innovation  req");
         require(techs.spacetimeWarp >= 3, "Level 3 Spacetime Warp req");
     }
 
-    function techCost(uint id) internal pure returns (S.ERC20s memory) {
+    function techCost(uint256 id) internal pure returns (S.ERC20s memory) {
         S.ERC20s memory cost;
         if (id == ID.ENERGY_INNOVATION) {
             cost.quartz = 800;
@@ -172,58 +145,20 @@ contract Lab {
         return cost;
     }
 
-    function _techsCost(
-        S.Techs memory techs
-    ) internal pure returns (S.TechsCost memory) {
+    function _techsCost(S.Techs memory techs) internal pure returns (S.TechsCost memory) {
         S.TechsCost memory cost;
-        cost.energyInnovation = techUpgradeCost(
-            techs.energyInnovation,
-            techCost(ID.ENERGY_INNOVATION)
-        );
-        cost.digitalSystems = techUpgradeCost(
-            techs.digitalSystems,
-            techCost(ID.DIGITAL_SYSTEMS)
-        );
-        cost.beamTechnology = techUpgradeCost(
-            techs.beamTechnology,
-            techCost(ID.BEAM_TECHNOLOGY)
-        );
-        cost.armourInnovation = techUpgradeCost(
-            techs.armourInnovation,
-            techCost(ID.ARMOUR_INNOVATION)
-        );
-        cost.ionSystems = techUpgradeCost(
-            techs.ionSystems,
-            techCost(ID.ION_SYSTEMS)
-        );
-        cost.plasmaEngineering = techUpgradeCost(
-            techs.plasmaEngineering,
-            techCost(ID.PLASMA_ENGINEERING)
-        );
-        cost.armsDevelopment = techUpgradeCost(
-            techs.armsDevelopment,
-            techCost(ID.ARMS_DEVELOPMENT)
-        );
-        cost.shieldTech = techUpgradeCost(
-            techs.shieldTech,
-            techCost(ID.SHIELD_TECH)
-        );
-        cost.spacetimeWarp = techUpgradeCost(
-            techs.spacetimeWarp,
-            techCost(ID.SPACETIME_WARP)
-        );
-        cost.combustiveDrive = techUpgradeCost(
-            techs.combustiveDrive,
-            techCost(ID.COMBUSTIVE_DRIVE)
-        );
-        cost.thrustPropulsion = techUpgradeCost(
-            techs.thrustPropulsion,
-            techCost(ID.THRUST_PROPULSION)
-        );
-        cost.warpDrive = techUpgradeCost(
-            techs.warpDrive,
-            techCost(ID.WARP_DRIVE)
-        );
+        cost.energyInnovation = techUpgradeCost(techs.energyInnovation, techCost(ID.ENERGY_INNOVATION));
+        cost.digitalSystems = techUpgradeCost(techs.digitalSystems, techCost(ID.DIGITAL_SYSTEMS));
+        cost.beamTechnology = techUpgradeCost(techs.beamTechnology, techCost(ID.BEAM_TECHNOLOGY));
+        cost.armourInnovation = techUpgradeCost(techs.armourInnovation, techCost(ID.ARMOUR_INNOVATION));
+        cost.ionSystems = techUpgradeCost(techs.ionSystems, techCost(ID.ION_SYSTEMS));
+        cost.plasmaEngineering = techUpgradeCost(techs.plasmaEngineering, techCost(ID.PLASMA_ENGINEERING));
+        cost.armsDevelopment = techUpgradeCost(techs.armsDevelopment, techCost(ID.ARMS_DEVELOPMENT));
+        cost.shieldTech = techUpgradeCost(techs.shieldTech, techCost(ID.SHIELD_TECH));
+        cost.spacetimeWarp = techUpgradeCost(techs.spacetimeWarp, techCost(ID.SPACETIME_WARP));
+        cost.combustiveDrive = techUpgradeCost(techs.combustiveDrive, techCost(ID.COMBUSTIVE_DRIVE));
+        cost.thrustPropulsion = techUpgradeCost(techs.thrustPropulsion, techCost(ID.THRUST_PROPULSION));
+        cost.warpDrive = techUpgradeCost(techs.warpDrive, techCost(ID.WARP_DRIVE));
         return cost;
     }
 }
